refactor auth and keyword apis
